Werror fixes
A bunch of fixes to get `-Werror` builds to pass,
both on Linux and on Windows (`/WX`)

* Disable verbose Windows warnings
  * `_CRT_SECURE_NO_WARNINGS` because of `std::getenv`
  * C4267 because of conversions from `size_t` to other integers
* Define `WIN32_LEAN_AND_MEAN` and `NOMINMAX` on Windows
  * Gets rid of some errors and speeds up the build
* Convert integer CUDA objects to std::uintptr_t
  before reinterpreting them to a pointer
* Fixed "unused function" warning for `GetHipFormatPixelSize`
  * There was a lot of duplication, now a single function
    called `imageElementByteSize`
* Mark some variables as potentially unused (only used in asserts)
* Other minor fixes

+++ source/adapters/cuda/sampler.cpp
@@ -18,7 +18,7 @@ urSamplerCreate(ur_context_handle_t hContext, const ur_sampler_desc_t *pDesc,
       new ur_sampler_handle_t_(hContext)};
 
   if (pDesc->stype == UR_STRUCTURE_TYPE_SAMPLER_DESC) {
-    Sampler->Props |= pDesc->normalizedCoords;
+    Sampler->Props |= static_cast<uint32_t>(pDesc->normalizedCoords);
     Sampler->Props |= pDesc->filterMode << 1;
     Sampler->Props |= pDesc->addressingMode << 2;
   } else {
